+++ packages/drag-n-drop-upload/CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## To be released
 
+- add support for a custom upload function in the drag-n-drop-upload plugin.
+
 ## 4.1.1
 
 - add `sideEffects` for css files [#1833](https://github.com/draft-js-plugins/draft-js-plugins/issues/1833)